QtPropertyBrowser: Respond to each change of text in QtLineEditFactory
I've extended this editor to allow adding a QCompleter to the line edit,
but when changing the value via the QCompleter only the 'textChanged'
and not the 'textEdited' signal is emitted.

'blockSignals' is now used to avoid this signal when applying a change
of property value, like for all other editors.

Change-Id: I1668dd45ad2e2bc651777b97cb143d1525eb5bd0
Reviewed-by: Jarek Kobus <jaroslaw.kobus@digia.com>