QtSingleApplication: Bail out of read loop when peer disconnects
When messages are received while the event loop is inactive
(for example, when locked by a native dialog), it can happen
that a new connection is triggered which then goes into
unconnected state when the peer terminates. Add a check
to the loop and discard the connection in that case.

Task-number: QTSOLBUG-193
Change-Id: I7d5502cb411d2b967bd06fe8734b97fee54b7d18
Reviewed-by: Joni Poikelin <joni.poikelin@qt.io>
Reviewed-by: Andy Shaw <andy.shaw@qt.io>